# Optimize BuildBidderNameHashSet