Pprof identified big amount of allocations caused by each call to `CoreBidderNames` on each auction request
By storing result of `CoreBidderNames` into a single variable number of allocations reduced 4x

` go test -benchmem -v -bench=Bench ./openrtb_ext/. |prettybench`

Before:
```
BenchmarkBuildBidderNameHashSet
PASS
benchmark                             iter      time/iter   bytes alloc         allocs
---------                             ----      ---------   -----------         ------
BenchmarkBuildBidderNameHashSet-18   50244    22.66 μs/op    21333 B/op   12 allocs/op
ok      github.com/prebid/prebid-server/v2/openrtb_ext  1.415s
```

After:
```
BenchmarkBuildBidderNameHashSet
PASS
benchmark                              iter       time/iter   bytes alloc        allocs
---------                              ----       ---------   -----------        ------
BenchmarkBuildBidderNameHashSet-18   152774   7813.00 ns/op    10312 B/op   3 allocs/op
ok      github.com/prebid/prebid-server/v2/openrtb_ext  1.309s
```