@SyntaxNode 2024-10-30
The `CoreBidders` slice is changed during startup when new aliases are configured. Could you please verify `BuildBidderNameHashSet` is only called after all `SetAliasBidderName` calls are made?