@SyntaxNode 2024-10-30
The `coreBidderNames` slice can be altered by calls to `SetAliasBidderName` during startup. Could you please confirm all calls to the optimized `BuildBidderNameHashSet` occur after any potenital call to `SetAliasBidderName` ?