@linux019 2024-10-31
main() calls `LoadBidderInfoFromDisk`
https://github.com/prebid/prebid-server/blob/5b11f59bdde56d7b22678efb4b3dfb2da9c62a7f/main.go#L35
Next is creates a HTTP server where a call to auction endpoint triggers `BuildBidderNameHashSet`
Yes if new changes will call `SetAliasBidderName` this cached set needs to be rebuilt.